LayoutTests/fast/dom/DOMURL/searchparams-expected.txt

CONSOLE MESSAGE: PASS
CONSOLE MESSAGE: PASS
CONSOLE MESSAGE: PASS
CONSOLE MESSAGE: PASS
CONSOLE MESSAGE: PASS threw exception TypeError: Cannot convert a symbol to a string.
